Normalize MFSDP data-parallel axes at fully_shard
